+++ crates/pdk/Cargo.toml
@@ -13,14 +13,10 @@ warpgate_pdk = { version = "0.14.1", path = "../warpgate-pdk" }
 extism-pdk = { workspace = true }
 rustc-hash = { workspace = true }
 serde = { workspace = true }
-tracing = { workspace = true, optional = true }
-tracing-subscriber = { workspace = true, optional = true, features = [
-	"registry",
-] }
 
 [dev-dependencies]
 proto_pdk = { path = ".", features = ["tracing"] }
 
 [features]
 default = []
-tracing = ["dep:tracing", "dep:tracing-subscriber"]
+tracing = ["warpgate_pdk/tracing"]
