+++ crates/pdk/src/lib.rs
@@ -1,10 +1,6 @@
 mod helpers;
 mod macros;
-#[cfg(feature = "tracing")]
-mod tracing;
 
 pub use helpers::*;
 pub use proto_pdk_api::*;
-#[cfg(feature = "tracing")]
-pub use tracing::*;
 pub use warpgate_pdk::*;
