+++ crates/warpgate-pdk/Cargo.toml
@@ -11,3 +11,14 @@ repository = "https://github.com/moonrepo/proto"
 warpgate_api = { version = "0.15.2", path = "../warpgate-api" }
 extism-pdk = { workspace = true }
 serde = { workspace = true }
+tracing = { workspace = true, optional = true }
+tracing-subscriber = { workspace = true, optional = true, features = [
+	"registry",
+] }
+
+[dev-dependencies]
+warpgate_pdk = { path = ".", features = ["tracing"] }
+
+[features]
+default = []
+tracing = ["dep:tracing", "dep:tracing-subscriber"]
