+++ crates/warpgate-pdk/src/lib.rs
@@ -1,7 +1,11 @@
 mod api;
 mod funcs;
 mod macros;
+#[cfg(feature = "tracing")]
+mod tracing;
 
 pub use api::*;
 pub use funcs::*;
+#[cfg(feature = "tracing")]
+pub use tracing::*;
 pub use warpgate_api::*;
